extra sprintf into own module
